Move sub-classes of DataGridViewRow class to separate file
